# Optimize replaceShortcodeTokens